added more NIRSpec nonlinear tests
